+++ backend/src/docs/asciidoc/index.adoc
@@ -13,5 +13,5 @@ include::pin.adoc[]
 include::atlas.adoc[]
 include::member.adoc[]
 include::permission.adoc[]
-include::oauth.adoc[]
+include::auth.adoc[]
 include::bookmark.adoc[]
