--- backend/src/docs/asciidoc/oauth.adoc
@@ -1,9 +1,0 @@
-== 소셜 로그인
-
-=== KAKAO 로그인 URL 반환
-
-operation::oauth-controller-test/redirection[snippets='http-request,http-response']
-
-=== KAKAO 로그인
-
-operation::oauth-controller-test/login[snippets='http-request,http-response']
