+++ backend/src/main/java/com/mapbefine/mapbefine/auth/dto/LoginTokens.java
@@ -1,0 +1,11 @@
+package com.mapbefine.mapbefine.auth.dto;
+
+public record LoginTokens(
+        String accessToken,
+        String refreshToken
+) {
+
+    public AccessToken toAccessToken() {
+        return new AccessToken(accessToken);
+    }
+}
