--- backend/src/main/java/com/mapbefine/mapbefine/oauth/dto/LoginTokens.java
@@ -1,14 +1,0 @@
-package com.mapbefine.mapbefine.oauth.dto;
-
-import com.mapbefine.mapbefine.member.dto.response.MemberDetailResponse;
-
-public record LoginTokens(
-        String accessToken,
-        String refreshToken,
-        MemberDetailResponse memberDetailResponse
-) {
-
-    public LoginInfoResponse toResponse() {
-        return new LoginInfoResponse(accessToken, memberDetailResponse);
-    }
-}
